update count to type to int

+++ api/src/main/java/io/opentelemetry/trace/Event.java
@@ -48,5 +48,5 @@ public interface Event {
    *
    * @return the number of dropped attributes.
    */
-  long getDroppedAttributeCount();
+  int getDroppedAttributeCount();
 }
